# Valhalla.Application/Requests/JwtAuthenticationRequest.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Valhalla.Application.Requests
{
    public class JwtAuthenticationRequest
    {
        public string Username { get; set; } = null!;
        public string Password { get; set; } = null!;
    }
}